backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
asgiref==3.11.1
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
